Add ability to load authenticated users from file Issue #3
We should consider having a way to update authenticated users while the server is live.
Perhaps have a timer that refreshes authUsers every minute in case someone changes it?
Should a button be placed on admin side to refresh from authorized users?
Should a WatchService be set up on the authorized.users file so that if the file changes, authorized users is updated?

The last option is the most slick, I think.