Links now work properly for linked files
Turns out the when the java WatchService detects a change in a soft or hard link it
appends a tile to the end of the file name, meaning changes for the
original file name weren't detected.

The server looks for changes with authorized.users and authorized.users~
and updates which users are authenticated accordingly.

Be aware that deployment should set restrictive permissions on the ~/server folder
created at production time.