Bring in photos #4 from dorfnerv2 This commit also contains extensive refactoring of the dorfner code to fit into the system and clean it a little, improve little details of UX.
# It does not work as intended at the moment:
You can only add photos for plants in bed 10 (for the moment).
Photos are not yet added to plant.component.
We should implement a search bar like in the search bar branch to look for the plant to add photo for instead of having to know the accession number.

Code from and credit to dorfnerv2-digital-display-garden
https://github.com/UMM-CSci-3601-S17/digital-display-garden-iteration-4-dorfner-v2.git